Add method to cancel bluetooth_gatt_start_notify upon ble device disconnect
Current bluetooth_gatt_start_notify would leak since there was
no way to cancel it when the device disconnected. This lead
to a build of of notify callbacks over time. Since all
listners have to reject each callback this made the code
slower and slower as each notify was built up until the
system became sluggish from 1000s of notify callbacks
that had to be enumerated and rejected

https://github.com/esphome/aioesphomeapi/pull/328 refactors
to prevent the linear searching of callbacks